analyzer: Add support for the Conda package manager
Implement support for the Python package manager Conda [1].
Code that this implementation shares with the existing Pip
implementation is moved to a `PythonSupport` class.

[1] https://docs.conda.io/en/latest/
