+++ .travis.yml
@@ -28,6 +28,7 @@ env:
     - ANDROID_SDK_VERSION="4333796"
     - BOWER_VERSION="1.8.8"
     - CONAN_VERSION="1.18.0"
+    - CONDA_VERSION="4.7.12.1"
     - FLUTTER_HOME="/opt/flutter"
     - FLUTTER_VERSION="v1.7.8+hotfix.3-stable"
     - GIMME_GO_VERSION="1.13" # Used internally by Travis.
@@ -76,6 +77,11 @@ install:
   - pip install --user conan==$CONAN_VERSION
   - conan user # Create the conan data directory. Automatic detection of your arch, compiler, etc.
   - pip install --user pipenv==$PYTHON_PIPENV_VERSION
+  - wget https://repo.continuum.io/miniconda/Miniconda3-${CONDA_VERSION}-Linux-x86_64.sh -O miniconda.sh;
+  - bash miniconda.sh -b -p $HOME/miniconda
+  - source "$HOME/miniconda/etc/profile.d/conda.sh"
+  - conda config --set always_yes yes --set changeps1 no
+  - conda update -q conda
 
 script:
   - set -o pipefail
