+++ analyzer/src/main/resources/META-INF/services/com.here.ort.analyzer.PackageManagerFactory
@@ -2,6 +2,7 @@ com.here.ort.analyzer.managers.Bower$Factory
 com.here.ort.analyzer.managers.Bundler$Factory
 com.here.ort.analyzer.managers.Cargo$Factory
 com.here.ort.analyzer.managers.Conan$Factory
+com.here.ort.analyzer.managers.Conda$Factory
 com.here.ort.analyzer.managers.DotNet$Factory
 com.here.ort.analyzer.managers.GoDep$Factory
 com.here.ort.analyzer.managers.GoMod$Factory
